Replace boot flow code with main tree drivers
The projects/ folder contained code which would override plat_*
functions to match the boot flow of a given target platform. At the
moment, this comes down to two different flows: Linux-like (FDT in first
kernel arg; QEMU, FVP, RPi) and Android-like (FDT compiled into
Hafnium). Get rid of the weak symbols and replace this with two
"drivers" and a `boot_flow` config in the board spec.

Bug: 117551352
Change-Id: I26176a4999a5c598eae9e542d635393ed5b0b840