VHE: Remove branch protection in secure_aem_v8a_fvp_vhe builds
If hafnium is built with branch protection enabled, the current mmu code
ensures that branch protection is enforced on all stage 1 translation
tables. This includes S-EL0 partitions that run on VHE. Removing branch
protection in the VHE builds means that we dont have to force all S-EL0
partitions to be built with VHE enabled. This will be done at a later
time.

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Change-Id: I3871fb75a676e41d9bddddf26ebeb3c7d10e241e