feat(gicr): introduce 'gicr_frames' for platforms
This patch introduces a new build variable to broadcast the
number of GICR frames supported by the platform. The GIC
initialization sequence relies on the LAST_FRAME bit from
the GICR and is not related to the number of CPUs on the
platform. Normally number of CPUs and GICR frames are equal,
but platforms with more GICR frames than number of CPUs are
not uncommon.

Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Change-Id: I3a2bdf9ab2d3300604355a59345105479630e66b