fix: increase heap size
Change [1] breaks one TF-A CI test with an out of memory error while
allocating page tables for accessing high address ranges. Provided
hypervisor memory pools are no longer defragmented this issue can now
happen more often and number of heap pages must be increased.
Change [2] is an out of tree change extending the static number of heap
pages for running Trusty on top of Hafnium. This is an opportunity to
incorporate this requirement.
Change [3] requires increasing heap size for each config to accomodate
the change where vCPUs are allocated from the heap as opposed to
statically during build time.

[1] https://review.trustedfirmware.org/c/hafnium/hafnium/+/16649
[2] https://gitlab.arm.com/arm-reference-solutions/build-scripts/-/blob/TC2-2022.08.12/files/hafnium/tc2/project/reference/0001-feat-tc-increase-secure_tc-heap-page-size.patch
[3] https://review.trustedfirmware.org/c/hafnium/hafnium/+/16778

Signed-off-by: Raghu Krishnamurthy <raghu.ncstate@gmail.com>
Signed-off-by: Olivier Deprez <olivier.deprez@arm.com>
Change-Id: Ida2b1172e8e6a17de6b854646a27d51da1616165